Extract TaskList interface
