+++ src/main/java/scottie/Scottie.java
@@ -7,6 +7,7 @@ import scottie.instructions.Instruction;
 import scottie.instructions.InvalidCommandException;
 import scottie.parser.Parser;
 import scottie.storage.FileStorage;
+import scottie.tasks.AutoSavingTaskList;
 import scottie.tasks.TaskList;
 import scottie.ui.Cli;
 import scottie.ui.Ui;
@@ -20,7 +21,7 @@ public class Scottie {
 
     private static final String INVALID_COMMAND_MESSAGE = "Um... sorry I don't know what \"%s\" means...";
 
-    private final TaskList taskList = new TaskList(new FileStorage(TASKS_DATA_FILE_PATH));
+    private final TaskList taskList = new AutoSavingTaskList(new FileStorage(TASKS_DATA_FILE_PATH));
 
     /**
      * Sends an input to Scottie to process and respond using the given Ui.
