+++ src/main/java/scottie/tasks/TaskList.java
@@ -1,55 +1,24 @@
 package scottie.tasks;
 
-import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.Comparator;
-import java.util.Iterator;
 import java.util.List;
-import java.util.stream.Collectors;
-
-import scottie.storage.Storage;
 
 /**
- * Encapsulates a list of Tasks.
- * The tasks are initially loaded from a Storage and any changes
- * to the list are always immediately saved to the Storage.
+ * API for a list of Tasks.
  */
-public class TaskList implements Iterable<Task> {
-    private final Storage storage;
-    private final List<Task> tasks;
-
-    /**
-     * Constructs a TaskList.
-     */
-    public TaskList(Storage storage) {
-        this.storage = storage;
-        this.tasks = new ArrayList<>();
-        for (String taskData : this.storage.loadData()) {
-            try {
-                this.addTask(Task.fromEncodedString(taskData));
-            } catch (InvalidTaskDataException e) {
-                e.printStackTrace();
-            }
-        }
-    }
-
+public interface TaskList extends Iterable<Task> {
     /**
      * Returns whether this TaskList has no Tasks.
      *
      * @return True if this TaskList has no Tasks, false otherwise.
      */
-    public boolean isEmpty() {
-        return this.tasks.isEmpty();
-    }
+    boolean isEmpty();
 
     /**
      * Returns the number of Tasks in this TaskList.
      *
-     * @return The number of Tasks in thie TaskList.
+     * @return The number of Tasks in this TaskList.
      */
-    public int size() {
-        return this.tasks.size();
-    }
+    int size();
 
     /**
      * Returns the Task at the given index.
@@ -58,9 +27,7 @@ public class TaskList implements Iterable<Task> {
      * @param index The index of the Task to return.
      * @return The Task at the specified index.
      */
-    public Task getTask(int index) {
-        return this.tasks.get(index);
-    }
+    Task getTask(int index);
 
     /**
      * Adds the given Task to the TaskList.
@@ -68,11 +35,7 @@ public class TaskList implements Iterable<Task> {
      *
      * @param task The Task to add the TaskList.
      */
-    public void addTask(Task task) {
-        this.tasks.add(task);
-        this.saveTasks();
-
-    }
+    void addTask(Task task);
 
     /**
      * Deletes the Task at the given index from the TaskList.
@@ -81,10 +44,7 @@ public class TaskList implements Iterable<Task> {
      *
      * @param index The index of the Task to remove from the list.
      */
-    public void deleteTask(int index) {
-        this.tasks.remove(index);
-        this.saveTasks();
-    }
+    void deleteTask(int index);
 
     /**
      * Returns whether the task at the given index is marked as done.
@@ -93,9 +53,7 @@ public class TaskList implements Iterable<Task> {
      * @param index The index of the Task to check.
      * @return Whether the task with the given index is done.
      */
-    public boolean isMarked(int index) {
-        return this.tasks.get(index).isMarked();
-    }
+    boolean isMarked(int index);
 
     /**
      * Marks the task at the given index as done.
@@ -103,10 +61,7 @@ public class TaskList implements Iterable<Task> {
      *
      * @param index The index of the Task to mark as done.
      */
-    public void markTask(int index) {
-        this.getTask(index).markAsDone();
-        this.saveTasks();
-    }
+    void markTask(int index);
 
     /**
      * Marks the task at the given index as not done.
@@ -114,10 +69,7 @@ public class TaskList implements Iterable<Task> {
      *
      * @param index The index of the Task to mark as not done.
      */
-    public void unmarkTask(int index) {
-        this.getTask(index).markAsUndone();
-        this.saveTasks();
-    }
+    void unmarkTask(int index);
 
     /**
      * Returns a List of Tasks in this TaskList whose
@@ -127,11 +79,7 @@ public class TaskList implements Iterable<Task> {
      * @param searchText The text to search for.
      * @return A List of matching Tasks.
      */
-    public List<Task> filterTasks(String searchText) {
-        return this.tasks.stream()
-                .filter(task -> task.matchesAgainst(searchText))
-                .collect(Collectors.toList());
-    }
+    List<Task> filterTasks(String searchText);
 
     /**
      * Sorts the Tasks in this TaskList by their descriptions.
@@ -139,14 +87,7 @@ public class TaskList implements Iterable<Task> {
      *
      * @param isReversed Whether to sort in the reversed order.
      */
-    public void sortByDescription(boolean isReversed) {
-        Comparator<Task> comparator = Comparator.comparing(Task::getDescription);
-        if (isReversed) {
-            comparator = comparator.reversed();
-        }
-        this.tasks.sort(comparator);
-        this.saveTasks();
-    }
+    void sortByDescription(boolean isReversed);
 
     /**
      * Sorts the Tasks in this TaskList by their descriptions.
@@ -155,26 +96,5 @@ public class TaskList implements Iterable<Task> {
      *
      * @param isReversed Whether to sort in the reversed order.
      */
-    public void sortByDate(boolean isReversed) {
-        Comparator<LocalDateTime> localDateTimeComparator =
-                isReversed ? Comparator.reverseOrder() : Comparator.naturalOrder();
-        this.tasks.sort(Comparator.comparing(Task::getDateTime,
-                Comparator.nullsLast(localDateTimeComparator)));
-        this.saveTasks();
-    }
-
-    /**
-     * Saves the tasks in this list to the storage.
-     * This method is called whenever the tasks in the list
-     * are modified.
-     */
-    private void saveTasks() {
-        List<String> encodedTasks = this.tasks.stream().map(Task::toEncodedString).collect(Collectors.toList());
-        this.storage.saveData(encodedTasks);
-    }
-
-    @Override
-    public Iterator<Task> iterator() {
-        return tasks.iterator();
-    }
+    void sortByDate(boolean isReversed);
 }
